Add files via upload
The macro is to book an inspection on a website. The macro checks if the desired date of inspection is available every 5 minutes. If available, it books inspection and prefer afternoon inspection over morning inspection.
It fills in some information of the person booking the inspection for contact.
Hope to get a notification of email upon successful booking.